+++ app/app-api/src/main/kotlin/com/beproud/appapi/connect/v1/ConnectService.kt
@@ -3,8 +3,7 @@ package com.beproud.appapi.connect.v1
 import com.beproud.appapi.connect.v1.dto.ConnectRequest
 import com.beproud.appapi.connect.v1.dto.ConnectResponse
 import com.beproud.appapi.user.v1.UserDomainService
-import com.beproud.config.auth.JwtTokenProvider
-import com.beproud.domain.rds.user.UserRepository
+import com.beproud.auth.JwtTokenProvider
 import mu.KotlinLogging
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
