+++ app/app-api/src/main/resources/application.yml
@@ -8,17 +8,33 @@ spring:
         import: "classpath:application-rds.yml"
     profiles:
         group:
-            h2: h2-db, docker-redis
+            h2: h2-db
             docker: docker-db, docker-redis
     main:
         allow-bean-definition-overriding: true
-    cache:
-        type: redis
-    security:
-        oauth2:
-            resourceserver:
-                jwt:
-                    issuer-uri: "http://auth-server:9090"
+#    cache:
+#        type: redis
+
+springdoc:
+    api-docs:
+        #        path: /api-docs
+        #        version: openapi_3_1
+        groups:
+            enabled: true
+    swagger-ui:
+#        version: v1.0.0
+        path: /swagger-ui.html
+        tags-sorter: alpha
+        operations-sorter: method
+        groups-order: DESC
+        disable-swagger-default-url: true
+        display-request-duration: true
+    packages-to-scan: com.beproud.appapi
+    paths-to-match:
+        - /api/v1/**
+    show-actuator: true
+    default-consumes-media-type: application/json
+    default-produces-media-type: application/json
 app:
     jwt:
         secret-key: 123
